\name{uskew-package}
\alias{uskew-package}
\alias{uskew}
\docType{package}
\title{
Finite mixture of unrestricted multivariate skew t distributions  
}
\description{
Theis package implements an EM algorithm for 
fitting mixtures of unrestricted multivariate skew t (FM-MST) distributions.
Functions for random sample generation, discriminant analysis, 
and visualisation (in 2D and 3D) is also provided.
}
\details{
\tabular{ll}{
Package: \tab uskew (EMMIX-uskew)\cr
Type: \tab Package\cr
Version: \tab 0.11-3\cr
Date: \tab 2013-10-15\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}

}
\author{
S. Lee, G.J. McLachlan
}
\references{
Lee S, McLachlan G (2011). On the fitting of mixtures of multivariate skew t-distributions
via the EM algorithm. arXiv:1109.4706 [stat.ME]

Lee, S. and McLachlan, G.J. (2012) Finite mixtures of multivariate 
skew t-distributions: some recent and new results. 
Statistics and Computing. To appear
                                     
Lee, S. X. and McLachlan, G.J. (2013) EMMIX-uskew: an R package for 
fitting mixtures of multivariate skew t-distributions via the EM algorithm.
Journal of Statistical Software. To appear
}
\keyword{ package }
\seealso{
\code{\link{fmmst}}, \code{\link{dfmmst}}, \code{\link{rfmmst}}, \code{\link{fmmst.contour.3d}}
}
